Rename release check module to web and add analytics ping
